Refactorize benchmarking
