+++ benchmark/btb_benchmark/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+
+"""Top-level package for BTB Benchmark."""
+
+from btb_benchmark.main import run_benchmark
+
+__author__ = """MIT Data To AI Lab"""
+__email__ = 'dailabmit@gmail.com'
+__version__ = '0.3.7.dev0'
+__all__ = ('run_benchmark', )
