+++ benchmark/btb_benchmark/challenges/bohachevsky.py
@@ -1,8 +1,7 @@
 # -*- coding: utf-8 -*-
 
 import numpy as np
-
-from btb.benchmark.challenges.challenge import Challenge
+from btb_benchmark.challenges.challenge import Challenge
 
 
 class Bohachevsky(Challenge):
