+++ benchmark/btb_benchmark/challenges/randomforest.py
@@ -1,9 +1,8 @@
 import logging
 
+from btb_benchmark.challenges.challenge import MLChallenge
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.metrics import f1_score
-
-from btb.benchmark.challenges.challenge import MLChallenge
 
 LOGGER = logging.getLogger(__name__)
