+++ benchmark/btb_benchmark/challenges/xgboost.py
@@ -1,9 +1,8 @@
 import logging
 
+from btb_benchmark.challenges.challenge import MLChallenge
 from sklearn.metrics import f1_score
 from xgboost import XGBClassifier
-
-from btb.benchmark.challenges.challenge import MLChallenge
 
 LOGGER = logging.getLogger(__name__)
